requirements.txt
web3>=6.0.0
torch>=2.0.0
tensorflow>=2.13.0
fastapi>=0.100.0
uvicorn>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
flask>=2.3.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
aiohttp>=3.8.0
web3[tester]>=6.0.0
eth-account>=0.9.0
eth-abi>=4.0.0
joblib>=1.3.0
protobuf>=4.24.0
